FEAT: Split some files, more refactor
