+++ cli/cli_handler.h
@@ -1,5 +1,5 @@
-#ifndef COMMAND_HANDLER_H
-#define COMMAND_HANDLER_H
+#ifndef ERD_CLI_HANDLER_H
+#define ERD_CLI_HANDLER_H
 
 #include "inputhandler/ext.h"
