+++ cli/inputhandler/ext.h
@@ -1,6 +1,5 @@
-#ifndef EXTINPUTHANDLER_H
-#define EXTINPUTHANDLER_H
-
+#ifndef ERD_CLI_INPUT_HANDLER_EXT_H
+#define ERD_CLI_INPUT_HANDLER_EXT_H
 
 #include "arghandler.h"
 #include "wrappers/transaction_input_wrapper.h"
