+++ tests/test_cli/test_inputhandler/CMakeLists.txt
@@ -3,6 +3,7 @@ include_directories(${PROJECT_SOURCE_DIR}/src)
 
 add_executable(test_inputhandler
 		${PROJECT_SOURCE_DIR}/cli/cli_handler.cpp
+		${PROJECT_SOURCE_DIR}/cli/cli_utility.cpp
 		${PROJECT_SOURCE_DIR}/cli/secretkeyprovider.cpp
 		${PROJECT_SOURCE_DIR}/cli/config/cliconfig.cpp
 		test_inputhandler.cpp)
